refactor(web): move species reconcile into species_catalog pkg
reconcile.py + shim species_catalog_reconcile_service. Refs #344 phase B.

Made-with: Cursor
