+++ app/web/services/species_catalog/__init__.py
@@ -18,6 +18,7 @@ from services.species_catalog.allowlist import (
     species_matches_allowlist,
     species_name_match_norm_keys,
 )
+from services.species_catalog.reconcile import reconcile_species_catalog
 
 __all__ = [
     "allowlist_scientific_name_for_display_name",
@@ -25,6 +26,7 @@ __all__ = [
     "ingest_name_matches_allowlist",
     "load_catalog_allowlist_names",
     "load_catalog_allowlist_norm_keys",
+    "reconcile_species_catalog",
     "resolve_allowlist_path",
     "species_matches_allowlist",
     "species_name_match_norm_keys",
