+++ docs/project/WEB_SERVICES_DOMAIN_MAP.md
@@ -10,7 +10,7 @@
 
 ## 1. `species_catalog`
 
-Подпакет кода: `app/web/services/species_catalog/` (фаза B [#344](https://github.com/Gfermoto/BirdLense-Hub/issues/344)) — первый модуль **`allowlist.py`**; корневой shim **`species_catalog_allowlist_service.py`** сохраняет старые импорты.
+Подпакет кода: `app/web/services/species_catalog/` (фаза B [#344](https://github.com/Gfermoto/BirdLense-Hub/issues/344)) — модули **`allowlist.py`**, **`reconcile.py`**; shims **`species_catalog_allowlist_service.py`**, **`species_catalog_reconcile_service.py`** сохраняют старые импорты.
 
 | Префикс / зона | Файлы маршрутов | Сервисы (ориентир) |
 |----------------|-----------------|---------------------|
